+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [Unreleased]
+### Added
+- Base64 and Base32 encoding/decoding utilities with secure_buffer overloads.
+
 ## [0.3.0] - 2025-09-05
 ### Added
 - PBKDF2 implementation.